SCUMM: HE: Fix TMSK ZPlanes handling
This closes ticket 13420:
"SCUMM: PUTTZOO Animation overlap bug
with final iceberg in Putt-Putt Saves the Zoo"

Not only this now works correctly, but now in updated versions
of the game the hair of the lower left penguin is not masked
anymore, so the iceberg piece will be drawn on top of said hair.
I have verified that this is 100% accurate, the code for TMSK
decompression is the same in every HE version, and the original
executable shows exactly the same behavior.

So refrain from filing a bug for that, thanks :P