SCUMM: HE/FREDDI3: Provide better solution for bug #13532
"SCUMM/HE: Flashing line at bottom of screen in Freddi 3"

This solves #13532 in a better way: instead of moving the
animation down by one pixel, we simply avoid marking that
line of stray pixels as part of the dirty rectangle.

We remove two pixels since AkosRenderer::paintCelByleRLE()
is going to add one pixel to the rect during the clipping part.