SCI: (CGA/Hercules) - add argument for cropped Hercules image
This is an idea that came up in the PR. It will remove the black border in
Hercules mode. Currently, there is no other way to enable it except in the
code. It could be added as a launcher option, but having a checkbox just
for the case where Hercules mode is enabled might be a bit too much.

I also turned the monochrome colors into an argument to make it more
flexible.